move our non-trivial implementations to another file so that it's easy to swap them for a different implementation
